Rename libratom.utils libratom.lib
