+++ libratom/cli/cli.py
@@ -13,7 +13,7 @@ import click_log
 import libratom.cli.subcommands as subcommands
 from libratom.cli import CONTEXT_SETTINGS, INT_METAVAR, PATH_METAVAR
 from libratom.cli.utils import PathPath, validate_out_path
-from libratom.utils.entities import SPACY_MODEL_NAMES, SPACY_MODELS
+from libratom.lib.entities import SPACY_MODEL_NAMES, SPACY_MODELS
 
 logger = logging.getLogger(__name__)
