+++ tests/unit/test_cli.py
@@ -10,13 +10,13 @@ from sqlalchemy.orm import sessionmaker
 
 import libratom
 from libratom.cli.cli import ratom
-from libratom.models.entity import Entity
-from libratom.utils.database import db_session
-from libratom.utils.entities import (
+from libratom.lib.database import db_session
+from libratom.lib.entities import (
     count_messages_in_files,
     extract_entities,
     process_message,
 )
+from libratom.models.entity import Entity
 
 
 class Expected:
